# doc/openapi.yml
openapi: "3.1.0"
info:
  title: OAuth2 + OIDC Authorization Server API
  version: "1.0.0"
  description: |
    自架 Authorization Server，支援 Authorization Code + PKCE、Refresh Token Rotation、
    Social Login（Google、Microsoft、Line、Facebook、Instagram、Threads）。

servers:
  - url: https://localhost:7001
    description: 本地開發

paths:
  /api/v1/account/register:
    post:
      summary: 本地帳密註冊
      tags: [Account]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        "201":
          description: 註冊成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        "400":
          description: 驗證失敗（密碼格式錯誤等）
        "409":
          description: Email 已被使用

  /api/v1/account/me:
    get:
      summary: 取得目前登入使用者資料
      tags: [Account]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 使用者資料
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        "401":
          description: 未授權

  /api/v1/account/external-logins:
    get:
      summary: 取得已綁定的 Social Provider 清單
      tags: [Account]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 綁定清單
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExternalLogin'

  /api/v1/account/external-logins/{provider}:
    delete:
      summary: 解除 Social Provider 綁定
      tags: [Account]
      security:
        - bearerAuth: []
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [Google, Microsoft, Facebook, Line, Threads]
      responses:
        "204":
          description: 解除綁定成功
        "400":
          description: 必須保留至少一種登入方式
        "404":
          description: 找不到此 Provider 的綁定

  /api/v1/account/password:
    put:
      summary: 修改密碼
      tags: [Account]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        "204":
          description: 修改成功
        "400":
          description: 原密碼錯誤或新密碼格式不符

  /connect/authorize:
    get:
      summary: Authorization Endpoint（OIDC）
      tags: [OIDC]
      parameters:
        - name: response_type
          in: query
          required: true
          schema:
            type: string
            enum: [code]
        - name: client_id
          in: query
          required: true
          schema:
            type: string
        - name: redirect_uri
          in: query
          required: true
          schema:
            type: string
            format: uri
        - name: scope
          in: query
          schema:
            type: string
        - name: code_challenge
          in: query
          required: true
          schema:
            type: string
        - name: code_challenge_method
          in: query
          required: true
          schema:
            type: string
            enum: [S256]
      responses:
        "302":
          description: 重導至登入頁或 redirect_uri

  /connect/token:
    post:
      summary: Token Endpoint（OIDC）
      tags: [OIDC]
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                grant_type:
                  type: string
                  enum: [authorization_code, refresh_token, client_credentials]
                code:
                  type: string
                redirect_uri:
                  type: string
                client_id:
                  type: string
                client_secret:
                  type: string
                code_verifier:
                  type: string
                refresh_token:
                  type: string
      responses:
        "200":
          description: Token 回應
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        "400":
          description: 錯誤（invalid_grant、invalid_client 等）

  /connect/userinfo:
    get:
      summary: UserInfo Endpoint（OIDC）
      tags: [OIDC]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 使用者資訊
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        "401":
          description: 未授權

  /connect/endsession:
    get:
      summary: End Session Endpoint（OIDC Logout）
      tags: [OIDC]
      parameters:
        - name: post_logout_redirect_uri
          in: query
          schema:
            type: string
            format: uri
        - name: id_token_hint
          in: query
          schema:
            type: string
      responses:
        "302":
          description: 登出並重導

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        displayName:
          type: string
          maxLength: 100

    RegisterResponse:
      type: object
      properties:
        userId:
          type: string
        email:
          type: string

    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string

    ExternalLogin:
      type: object
      properties:
        loginProvider:
          type: string
        providerDisplayName:
          type: string

    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string

    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
        expires_in:
          type: integer
        refresh_token:
          type: string
        id_token:
          type: string
        scope:
          type: string

    UserInfo:
      type: object
      properties:
        sub:
          type: string
        email:
          type: string
        name:
          type: string
        picture:
          type: string